other_2_webscrapping_airflow/requirements.txt
selenium
webdriver-manager
pandas
mysql-connector
gcloud